Cart.jsの内容をApp.jsにうつす
