+++ src/App.scss
@@ -162,3 +162,71 @@ nav {
     }
   }
 }
+
+.emptyMessage {
+  text-transform: lowercase;
+  color: #617d98;
+  margin-top: 1rem;
+  text-align: center;
+}
+
+.cartItem {
+  display: grid;
+  align-items: center;
+  grid-template-columns: auto 1fr auto;
+  grid-column-gap: 1.5rem;
+  margin: 1.5rem 0;
+
+  img {
+    width: 5rem;
+    height: 5rem;
+    object-fit: cover;
+  }
+
+  .item-info {
+    h4 {
+      margin-bottom: 0.5rem;
+      font-weight: 500;
+      letter-spacing: 2px;
+
+      .price {
+        color: #617d98;
+      }
+    }
+
+    .remove-btn {
+      color: #2680c0;
+      letter-spacing: 0.25rem;
+      cursor: pointer;
+      font-size: 0.85rem;
+      background: transparent;
+      border: none;
+      margin-top: 0.375rem;
+      transition: all 0.3s linear;
+    }
+    .remove-btn:hover {
+      color: hsl(205, 84%, 74%);
+    }
+  }
+
+  .amount-container {
+    .amount-btn {
+      width: 1.5rem;
+      background: transparent;
+      border: none;
+      cursor: pointer;
+      color: #2680c0;
+    }
+
+    .amount-btn:hover {
+      color: hsl(205, 84%, 74%);
+    }
+
+    .amount {
+      text-align: center;
+      margin-bottom: 0;
+      font-size: 1.25rem;
+      line-height: 1;
+    }
+  }
+}
